data/data_api/requirements.txt
google-cloud-bigquery
spotipy
pandas
requests
python-dotenv
